+++ src/app/header/header.component.html
@@ -20,14 +20,10 @@
     </div>
 
     <div class="button-container">
-      <button
-        mat-raised-button
-        color="primary"
-        routerLink="/login"
-        class="loginButton"
-      >
-      null
+      <button mat-raised-button color="primary" routerLink="/login" class="loginButton">
+        {{ selectedPersonName || 'Login' }}
       </button>
+      
     </div>
   </div>
 </div>
