+++ src/app/header/header.component.ts
@@ -1,15 +1,26 @@
-import { Component, Inject, Input } from '@angular/core';
-import { PersonService } from 'src/app/services/person.service';
-import { LoginBoxComponent } from '../Components/login-box/login-box.component';
-
+import { Component, OnInit } from '@angular/core';
+import { PersonService, SelectedPersonService } from 'src/app/services/person.service';
 
 @Component({
   selector: 'app-header',
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss']
 })
-export class HeaderComponent {
+export class HeaderComponent implements OnInit {
+  selectedPersonName: string = '';
 
+  constructor(
+    private personService: PersonService,
+    private selectedPersonService: SelectedPersonService
+  ) {}
 
- 
+  ngOnInit(): void {
+    this.selectedPersonService.getPersonId().subscribe(personId => {
+      if (personId) {
+        this.personService.getPerson(+personId).subscribe(person => {
+          this.selectedPersonName = person.name;
+        });
+      }
+    });
+  }
 }
